+++ lib/authrocket/app_hook.rb
@@ -17,6 +17,7 @@ module AuthRocket
           org.created  org.updated  org.deleted
           realm.created  realm.updated  realm.deleted
           user.created  user.updated  user.deleted
+            user.email.verification_requested  user.email.verified
             user.login.succeeded  user.login.failed
             user.password_token.created  user.password_token.consumed  user.password_token.failed
         ).sort